overlays: Extract modal methods in separate file.
