+++ web/src/overlays_modals.ts
@@ -1,5 +1,6 @@
-import * as overlays from './overlays';
+import * as modals from "./modals";
+import * as overlays from "./overlays";
 
 export function any_active(): boolean {
-    return overlays.is_active() || overlays.is_modal_open();
+    return overlays.is_active() || modals.is_modal_open();
 }
